requirements.txt
Django==1.5.2
Markdown==2.3.1
South==0.8.2
argparse==1.2.1
django-allauth==0.12.0
django-filter==0.7
djangorestframework==2.3.7
oauthlib==0.5.1
psycopg2==2.5.1
python-openid==2.2.5
requests==1.2.3
requests-oauthlib==0.3.3
Pillow==2.1.0
boto==2.10.0
django-allauth==0.12.0
django-filter==0.7
django-storages==1.1.8
